max agents check

+++ api/config.go
@@ -14,9 +14,9 @@ type Config struct {
 	Domains       []string `yaml:"domains"`
 	Address       string   `yaml:"address"`
 	ReqMaxSize    int64    `yaml:"req_max_size"`
-	TokenTTL      int64    `yaml:"token_ttl"`
+	TokenTTL      int      `yaml:"token_ttl"`
 	Secret        string   `yaml:"secret"`
-	MaxAgents     int64    `yaml:"max_agents_per_user"` // TODO: implement check
-	CacheTTL      int64    `yaml:"cache_ttl"`
+	MaxAgents     int      `yaml:"max_agents_per_user"`
+	CacheTTL      int      `yaml:"cache_ttl"`
 	AllowNewUsers bool     `yaml:"allow_new_users"`
 }
